+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Make sure we shard alloy-metrics based on all metrics for all tenants and not for the giantswarm tenant only.
+
 ## [0.22.1] - 2025-03-25
 
 ### Fixed