+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add multi-tenancy support to alerting and recording rules loading by setting up the alloy-rules config.
 
+### Fixed
+
+- Make sure we shard alloy-metrics based on all metrics for all tenants and not for the giantswarm tenant only.
+
 ## [0.22.1] - 2025-03-25
 
 ### Fixed