+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.23.0] - 2025-04-07
+
 ### Added
 
 - Add multi-tenancy support to alerting and recording rules loading by setting up the alloy-rules config.
@@ -396,7 +398,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initialize project and create heartbeat for the installation.
 
-[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.22.1...HEAD
+[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.23.0...HEAD
+[0.23.0]: https://github.com/giantswarm/observability-operator/compare/v0.22.1...v0.23.0
 [0.22.1]: https://github.com/giantswarm/observability-operator/compare/v0.22.0...v0.22.1
 [0.22.0]: https://github.com/giantswarm/observability-operator/compare/v0.21.1...v0.22.0
 [0.21.1]: https://github.com/giantswarm/observability-operator/compare/v0.21.0...v0.21.1